requirements.txt
# Telegram Bot
python-telegram-bot>=20.0
httpx>=0.28.0

# Scheduling
apscheduler>=3.10.0
pytz>=2023.3

# Data Processing
pandas>=1.5.0
numpy>=1.23.0
python-dateutil>=2.8.2

# API Requests
requests>=2.28.0

# Data Visualization
plotly>=5.0.0
matplotlib>=3.7.0

# Machine Learning
scikit-learn>=1.2.0
stable-baselines3>=2.0.0
gymnasium>=0.28.0
xgboost>=2.0.0
torch>=2.0.0

# Macroeconomic Data (FRED API)
fredapi>=0.5.0

# Environment Variables
python-dotenv>=1.0.0

# News + Sentiment (GNews + FinBERT)
gnews>=0.3.0
transformers>=4.37.0
newspaper3k>=0.2.8
beautifulsoup4>=4.12.0
lxml>=4.9.0
lxml_html_clean>=0.1.1

# Gemini (new SDK)
google-genai>=0.5.0
